price-service/requirements.txt
mysql-connector
requests
apscheduler